Deprecate `devices` command in favor of `device list`
Change-type: minor
