+++ src/help.ts
@@ -231,7 +231,6 @@ See: https://git.io/JRHUW#deprecation-policy`,
 		'ssh',
 		'fleets',
 		'fleet',
-		'devices',
 		'device',
 		'tunnel',
 		'preload',
